requirements.txt
playwright>=1.40
gspread>=6.0
google-auth>=2.0
pyyaml>=6.0
tenacity>=8.0
click>=8.0
